pc配置中心-社区配置100%

+++ core/src/main/java/com/smart/home/modules/community/service/CommunityService.java
@@ -1,15 +1,22 @@
 package com.smart.home.modules.community.service;
 
 import com.github.pagehelper.PageHelper;
+import com.smart.home.common.exception.DuplicateDataException;
+import com.smart.home.common.exception.ServiceException;
+import com.smart.home.modules.community.dao.CommunityCategoryMapper;
 import com.smart.home.modules.community.dao.CommunityMapper;
+import com.smart.home.modules.community.dao.CommunityPostMapper;
 import com.smart.home.modules.community.entity.Community;
+import com.smart.home.modules.community.entity.CommunityCategory;
 import com.smart.home.modules.community.entity.CommunityExample;
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import javax.annotation.Resource;
 import java.util.List;
 import java.util.Date;
+import java.util.Objects;
 
 /**
  * @author jason
@@ -19,13 +26,38 @@ public class CommunityService {
 
     @Resource
     CommunityMapper communityMapper;
+    @Resource
+    CommunityCategoryMapper communityCategoryMapper;
+    @Resource
+    CommunityPostMapper communityPostMapper;
 
     public int create(Community community) {
+        CommunityCategory communityCategory = communityCategoryMapper.selectByPrimaryKey(community.getCategoryId().intValue());
+        if (Objects.isNull(communityCategory)) {
+            throw new ServiceException("社区类目不存在");
+        }
+        CommunityExample example = new CommunityExample();
+        example.createCriteria().andTitleEqualTo(community.getTitle());
+        if (communityMapper.countByExample(example) > 0) {
+            throw new DuplicateDataException("该社区已经存在");
+        }
+        community.setCategoryName(communityCategory.getTitle());
+        community.setFollowerCount(0);
+        community.setRevision(0);
         community.setCreatedTime(new Date());
         return communityMapper.insertSelective(community);
     }
 
     public int update(Community community) {
+        CommunityCategory communityCategory = communityCategoryMapper.selectByPrimaryKey(community.getCategoryId().intValue());
+        if (Objects.isNull(communityCategory)) {
+            throw new ServiceException("社区类目不存在");
+        }
+        CommunityExample example = new CommunityExample();
+        example.createCriteria().andTitleEqualTo(community.getTitle()).andIdNotEqualTo(community.getId());
+        if (communityMapper.countByExample(example) > 0) {
+            throw new DuplicateDataException("该社区已经存在");
+        }
         community.setUpdatedTime(new Date());
         return communityMapper.updateByPrimaryKeySelective(community);
     }
@@ -37,6 +69,10 @@ public class CommunityService {
     @Transactional(rollbackFor = RuntimeException.class)
     public void delete(List<Long> idList) {
         for (Long id : idList) {
+            // 检查有没有关联的帖子
+            if (communityPostMapper.countByCommunityId(id.intValue()) > 0) {
+                throw new ServiceException("已经有关联的帖子了，不能删除");
+            }
             communityMapper.deleteByPrimaryKey(id.intValue());
         }
     }
@@ -45,7 +81,13 @@ public class CommunityService {
         PageHelper.startPage(pageNum, pageSize);
         CommunityExample example = new CommunityExample();
         CommunityExample.Criteria criteria = example.createCriteria();
-        // TODO 按需根据字段查询
+        if (StringUtils.isNotBlank(community.getTitle())) {
+            criteria.andTitleEqualTo(community.getTitle());
+        }
+        if (community.getCategoryId() != null) {
+            criteria.andCategoryIdEqualTo(community.getCategoryId());
+        }
+        example.setOrderByClause("sort desc");
         return communityMapper.selectByExample(example);
     }
 
@@ -54,4 +96,9 @@ public class CommunityService {
         return community;
     }
 
+    public List<Community> queryAllValid() {
+        CommunityExample example = new CommunityExample();
+        example.setOrderByClause("sort desc");
+        return communityMapper.selectByExample(example);
+    }
 }
